+++ src/main/java/org/rumbledb/runtime/functions/util/formatting/namedtimezones/formatter/TimezonePlaceFormatter.java
@@ -1,4 +1,4 @@
-package org.rumbledb.runtime.functions.util.formatting.timezone.formatter;
+package org.rumbledb.runtime.functions.util.formatting.namedtimezones.formatter;
 
 import java.time.OffsetDateTime;
 import java.time.ZoneId;
